add tests for indexers
